requirements.txt
Flask==3.0.3
gunicorn==21.2.0

# Gemini
google-generativeai>=0.8.3

# Google API 
google-api-python-client==2.144.0
google-auth==2.35.0
google-auth-oauthlib==1.2.1
google-auth-httplib2==0.2.0

# 
protobuf>=4.25.3

# others
python-dateutil==2.9.0.post0
dateparser==1.2.0

pytz==2024.1
requests==2.32.3
google-cloud-discoveryengine==0.11.9
beautifulsoup4==4.12.2

# GCS for state files
google-cloud-storage==2.18.2